refactor: move top-level files to appropriate modules
- Move wut.py to gptme/cli/wut.py (CLI tool belongs with other CLI modules)
- Move _keyword_matching.py to gptme/util/keyword_matching.py (shared utility)
- Update pyproject.toml entrypoint for gptme-wut
- Update all import paths in affected files

Per Issue #1171 follow-up request from Erik.
